refactor: reorganize SetAPIsTemplate component and update imports for improved structure
